moved redux dir into src dir
