+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { Provider } from "react-redux";
 import { ExampleFeature } from "./domains/example-feature/example-feature.index";
-import { store } from "../redux/store";
+import { store } from "./redux/store";
 
 // This will be used to compose all the components from this feature
 // All routing will also live here
